requirements.txt
pyotp
qrcode
Pillow
